requirements.txt
aiodns==3.0.0
aiofiles==0.6.0
aiohttp==3.7.4.post0
aiosignal==1.2.0
aiosqlite==0.17.0
anyio==3.5.0
apischema==0.16.6
asgiref==3.5.0
async-generator==1.10
async-timeout==3.0.1
asyncio-extras==1.3.2
asyncpraw==7.5.0
asyncprawcore==2.3.0
attrs==21.4.0
authorizenet==1.1.4
autobahn==22.3.2
Automat==20.2.0
based58==0.1.0
beautifulsoup4==4.10.0
binance-connector==1.11.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
constantly==15.1.0
construct==2.10.67
construct-typing==0.5.2
cryptography==36.0.2
dateparser==1.1.1
discord.py==1.7.3
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.3
django-environ==0.9.0
django-phonenumber-field==6.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
Faker==13.3.5
filelock==3.7.1
frozenlist==1.3.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.7
httpx==0.18.2
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jsonrpcclient==4.0.2
jsonrpcserver==5.0.7
jsonschema==4.4.0
lxml==4.8.0
multidict==6.0.2
names==0.3.0
openpyxl==3.0.10
OSlash==0.6.3
paymongo==0.1.3.dev0
phonenumbers==8.12.52
Pillow==9.2.0
pipenv==2022.6.7
platformdirs==2.5.2
praw==7.6.0
prawcore==2.3.0
psaw==0.1.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.1.2
pycparser==2.21
PyJWT==2.4.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyrsistent==0.18.1
python-binance==1.0.15
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyXB==1.2.5
regex==2022.3.2
requests==2.27.1
rfc3986==1.5.0
schedule==1.1.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
solana==0.23.0
soupsieve==2.3.1
sqlparse==0.4.2
Twisted==22.2.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
types-cachetools==4.2.10
typing-extensions==3.10.0.2
tzdata==2022.1
tzlocal==4.1
ujson==5.1.0
update-checker==0.18.0
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
websocket-client==1.3.2
websockets==10.2
whitenoise==6.0.0
yarl==1.7.2
zope.interface==5.4.0
